feat: add export confirmation & update translations
- Add async confirmation dialog
- Add export/quit strings (EN/FR)
- Refactor export logic with switch expressions
- Replace old restart/quit commands by export/quit
- Update and simplify EndExperimentPage UI
- Clean up code and namespaces
